backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
elasticsearch==8.11.0
hazelcast-python-client==5.3.0
pydantic==2.5.0
python-dotenv==1.0.0